SFEperl-test-pod: package for perl-522 is added and for perl-520 is obsolete
--HG--
branch : spec_SFEperl-test-pod

+++ specs/SFEperl-test-pod.spec
@@ -4,7 +4,7 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 0
 
 %define cpan_name Test-Pod
@@ -33,6 +33,8 @@ BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/pathtools-584
 BuildRequires:    library/perl-5/pod-simple-584
 BuildRequires:    library/perl-5/test-simple-584
+BuildRequires:    library/perl-5/pod-simple-584
+BuildRequires:    library/perl-5/test-simple-584
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-simple-584
@@ -49,6 +51,8 @@ Summary:          check for POD errors in files
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
 BuildRequires:    library/perl-5/pathtools-510
+BuildRequires:    library/perl-5/pod-simple-510
+BuildRequires:    library/perl-5/test-simple-510
 BuildRequires:    library/perl-5/pod-simple-510
 BuildRequires:    library/perl-5/test-simple-510
 Requires:         runtime/perl-510 = *
@@ -69,6 +73,8 @@ BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/pathtools-512
 BuildRequires:    library/perl-5/pod-simple-512
 BuildRequires:    library/perl-5/test-simple-512
+BuildRequires:    library/perl-5/pod-simple-512
+BuildRequires:    library/perl-5/test-simple-512
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-simple-512
@@ -87,6 +93,9 @@ BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/pathtools-516
 BuildRequires:    library/perl-5/pod-simple-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/pod-simple-516
+BuildRequires:    library/perl-5/test-simple-516
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-simple-516
@@ -96,28 +105,30 @@ Requires:         library/perl-5/test-simple-516
 check for POD errors in files
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          check for POD errors in files
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/pathtools-520
-BuildRequires:    library/perl-5/pod-simple-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/pathtools-522
+BuildRequires:    library/perl-5/pod-simple-522
+BuildRequires:    library/perl-5/test-simple-522
+BuildRequires:    library/perl-5/pod-simple-522
+BuildRequires:    library/perl-5/test-simple-522
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/pod-simple-520
-Requires:         library/perl-5/test-simple-520
+Requires:         library/perl-5/pod-simple-522
+Requires:         library/perl-5/test-simple-522
 
-%description 520
+%description 522
 check for POD errors in files
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -130,8 +141,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -146,7 +166,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -163,7 +183,7 @@ modify_bin_dir() {
     then
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -220,8 +240,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -282,18 +302,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Sat Dec 05 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 1.51
 * Tue Dec 09 2014 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
